Update: Change New plugin structure - Egor
