+++ checkbox.php
@@ -1,5 +1,4 @@
 <?php
-
 /**
  * Plugin Name: WooCommerce Checkbox Integration
  * Plugin URI: https://morkva.co.ua/shop/checkbox-woocommerce?utm_source=checkbox-plugin
@@ -16,1625 +15,82 @@
  * WC tested up to: 7.3.0
  */
 
-if (! defined('ABSPATH')) {
+#  Stop access .php files through URL
+if (! defined('ABSPATH')) 
+{
     exit;
 }
 
-define('CHECKBOX_VERSION', '2.0.0');
+# Versions number
+define('CHECKBOX_VERSION', '1.2.0');
 define('CHECKBOX_LICENSE', 'free');
 
+# Include autoload
 require_once 'vendor/autoload.php';
-
-/**
- * INCLUDE CHECKBOX API LIBRARY
- */
+# Include checkbox api library
 require_once 'includes/class-api.php';
-
-/**
- * INCLUDE SUPPORT
- */
+# Include support
 require_once 'includes/class-support.php';
-
-/**
- * INCLUDE KLOGGER
- */
+# Include klogger
 require_once 'includes/class-klogger.php';
-
-// -----------------------------------------------------------------------//
-// --------------------------------SETUP----------------------------------//
-// -----------------------------------------------------------------------//
-
-add_action('admin_init', 'mrkv_checkbox_register_mysettings');
-if (! function_exists('mrkv_checkbox_register_mysettings')) {
-
-    /** Register plugin options */
-    function mrkv_checkbox_register_mysettings()
-    {
-        $options = array(
-            'ppo_login',
-            'ppo_password',
-            'ppo_cashier_name',
-            'ppo_cashier_surname',
-            'ppo_cashbox_key',
-            'ppo_tax_code',
-            'ppo_receipt_footer',
-            'ppo_autocreate',
-            'ppo_autocreate_receipt_order_statuses',
-            'ppo_autocreate_payment_order_statuses',
-            'ppo_payment_type',
-            'ppo_connected',
-            'ppo_autoopen_shift',
-            'ppo_sign_method',
-            'ppo_skip_receipt_creation',
-            'ppo_is_dev_mode',
-            'ppo_logger',
-            'ppo_cashbox_edrpou'
-        );
-        foreach ($options as $option) {
-            register_setting('ppo-settings-group', $option);
-        }
-    }
-}
-
-add_action('admin_menu', 'mrkv_checkbox_register_plugin_page_in_name');
-if (! function_exists('mrkv_checkbox_register_plugin_page_in_name')) {
-
-    /**
-     * Register plugin page in admin menu
-     */
-    function mrkv_checkbox_register_plugin_page_in_name()
-    {
-        add_menu_page(__('Налаштування пРРО Checkbox', 'checkbox'), __('Checkbox', 'checkbox'), 'manage_woocommerce', 'checkbox_settings', 'mrkv_checkbox_show_plugin_admin_page', plugin_dir_url(__FILE__) . 'assets/img/logo.svg');
-    }
-}
-
-/** Add action for plugin cron event 'checkbox_close_shift' */
-add_action('checkbox_close_shift', 'mrkv_checkbox_disconnect');
-/** Add action for plugin cron event 'checkbox_open_shift' */
-add_action('checkbox_open_shift', 'mrkv_checkbox_connect');
-
-// -----------------------------------------------------------------------//
-// -----------------------WOOCOMMERCE CUSTOMISATION-----------------------//
-// -----------------------------------------------------------------------//
-
-add_action('woocommerce_order_actions', 'mrkv_checkbox_wc_add_order_meta_box_action');
-if (! function_exists('mrkv_checkbox_wc_add_order_meta_box_action')) {
-
-    /**
-     * Add order metabox action
-     *
-     * @param array $actions all actions registered in order meta box
-     * @return array $actions all actions registered in order meta box
-     */
-    function mrkv_checkbox_wc_add_order_meta_box_action($actions)
-    {
-        $actions['create_bill_action'] = __('Створити чек', 'checkbox');
-        return $actions;
-    }
-}
-
-add_action('woocommerce_order_action_create_bill_action', 'mrkv_checkbox_wc_process_order_meta_box_action');
-if (! function_exists('mrkv_checkbox_wc_process_order_meta_box_action')) {
-    /**
-     * Process order metabox action
-     *
-     * @param WC_Order $order Order Info
-     */
-    function mrkv_checkbox_wc_process_order_meta_box_action($order)
-    {
-        $login       = get_option('ppo_login');
-        $password    = get_option('ppo_password');
-        $cashbox_key = get_option('ppo_cashbox_key');
-
-        if (! $login) {
-            $order->add_order_note(__('Вкажіть логін в налаштуваннях плагіна Checkbox', 'checkbox'), $is_customer_note = 0, $added_by_user = false);
-            return;
-        }
-        if (! $password) {
-            $order->add_order_note(__('Вкажіть пароль в налаштуваннях плагіна Checkbox', 'checkbox'), $is_customer_note = 0, $added_by_user = false);
-            return;
-        }
-        if (! $cashbox_key) {
-            $order->add_order_note(__('Вкажіть ліцензійний ключ віртуального касового апарату в налаштуваннях плагіна Checkbox', 'checkbox'), $is_customer_note = 0, $added_by_user = false);
-            return;
-        }
-
-        $is_dev = boolval(get_option('ppo_is_dev_mode'));
-        $logger = new Checkbox\KLoggerDecorator(boolval(get_option('ppo_logger')));
-        $api = new Checkbox\API($login, $password, $cashbox_key, $is_dev);
-
-        /** Check if receipt is already created */
-        if (! empty(get_post_meta($order->get_id(), 'receipt_id', true))) {
-            $order->add_order_note(__('Чек вже створено', 'checkbox'), $is_customer_note = 0, $added_by_user = false);
-            $logger->info(sprintf('Замовлення №%d. %s', $order->get_id(), __('Чек вже створено', 'checkbox')));
-            return;
-        }
-
-        /** Check current shift status */
-        $current_shift = $api->getCurrentCashierShift();
-        if (! isset($current_shift['status']) && ( 'OPENED' !== $current_shift['status'] )) {
-            /** Check if Autoopen shift feature is activated */
-            if(is_null(get_option('ppo_autoopen_shift'))){
-                update_option( 'ppo_autoopen_shift', 1 );
-            }
-            if (1 === (int) get_option('ppo_autoopen_shift')) {
-                mrkv_checkbox_connect();
-                sleep(8); // wait for 8 sec while shift is opening
-
-                if (0 === (int) get_option('ppo_connected')) {
-                    mrkv_checkbox_connect();
-                }
-            } else {
-                $order->add_order_note(__('Зміна не відкрита', 'checkbox'), $is_customer_note = 0, $added_by_user = false);
-                $logger->info(sprintf('Замовлення №%d. %s', $order->get_id(), __('Зміна не відкрита', 'checkbox')));
-                return;
-            }
-        }
-
-        $result = mrkv_checkbox_create_receipt($api, $order);
-
-        if ($result['success']) {
-            $order->add_order_note(sprintf('%s. <a href="%s" target="_blank">%s</a>', __('Чек створено', 'checkbox'), "https://check.checkbox.ua/{$result['receipt_id']}", __('Роздрукувати', 'checkbox')), $is_customer_note = 0, $added_by_user = false);
-
-            $logger->info(sprintf('Замовлення №%d. %s.', $order->get_id(), __('Чек створено', 'checkbox')));
-        } else {
-            $order->add_order_note(__('Виникла помилка під час створення чека', 'checkbox') . '.' . __('Повідомлення:', 'checkbox') . ' ' . $result['message'], $is_customer_note = 0, $added_by_user = false);
-
-            $logger->error(sprintf('Замовлення №%d. %s.', $order->get_id(), __('Виникла помилка під час створення чека', 'checkbox') . '.'
-            . __('Повідомлення:', 'checkbox') . ' ' . $result['message'] . '. '
-            . __('Деталі:', 'checkbox') . " loc:{$result['detail']['loc']}, msg:{$result['detail']['msg']}, type:{$result['detail']['type']}"));
-            $logger->debug(__('Зміст помилки:', 'checkbox'), $result);
-        }
-    }
-}
-
-add_filter('manage_edit-shop_order_columns', 'mrkv_checkbox_wc_new_order_column');
-if (! function_exists('mrkv_checkbox_wc_new_order_column')) {
-
-    /**
-     * Add order admin column
-     *
-     * @param array $columns Columns from edit shop page
-     * @return array $columns Updated columns from edit shop page
-     **/
-    function mrkv_checkbox_wc_new_order_column($columns)
-    {
-        $columns['receipt_column'] = __('ID Чека', 'checkbox');
-        return $columns;
-    }
-}
-
-add_action('manage_shop_order_posts_custom_column', 'mrkv_checkbox_wc_cogs_add_order_receipt_column_content');
-if (! function_exists('mrkv_checkbox_wc_cogs_add_order_receipt_column_content')) {
-
-    /**
-     * Fill ID Receipt column
-     *
-     * @param string $column column name
-     */
-    function mrkv_checkbox_wc_cogs_add_order_receipt_column_content($column)
-    {
-        global $the_order;
-
-        if ('receipt_column' === $column) {
-            $receipt_id = get_post_meta($the_order->get_id(), 'receipt_id', true);
-            printf('<a href="%s" target="_blank">%s</a>', "https://check.checkbox.ua/{$receipt_id}", $receipt_id);
-        }
-    }
-}
-
-// -----------------------------------------------------------------------//
-// -----------------------------CREATE RECEIPT----------------------------//
-// -----------------------------------------------------------------------//
-
-add_action('woocommerce_order_status_changed', 'mrkv_checkbox_auto_create_receipt', 99, 3);
-if (! function_exists('mrkv_checkbox_auto_create_receipt')) {
-    /**
-     * Automatic receipt creation
-     *
-     * @param string $order_id Order ID
-     * @param string $old_status old order status
-     * @param string $new_status new order status
-     */
-    function mrkv_checkbox_auto_create_receipt($order_id, $old_status, $new_status)
-    {
-
-        if (1 !== (int) get_option('ppo_autocreate')) {
-            return;
-        }
-
-        $order = wc_get_order($order_id);
-        $order_payment_id = $order->get_payment_method();
-
-        $order_statuses = (array) get_option('ppo_autocreate_receipt_order_statuses');
-        $payment_order_statuses = (array) get_option('ppo_autocreate_payment_order_statuses');
-
-        if (empty($order_statuses) && empty($payment_order_statuses)) {
-            return;
-        }
-
-        if (in_array($new_status, $order_statuses) || in_array($new_status, $payment_order_statuses[$order_payment_id])) {
-
-            /** Check whether to create receipt or not */
-            /*$ppo_skip_receipt_creation = get_option('ppo_skip_receipt_creation');
-            if (isset($ppo_skip_receipt_creation[ $order->get_payment_method() ]) && 'yes' === $ppo_skip_receipt_creation[ $order->get_payment_method() ]) {
-                $order->add_order_note(__('Чек не створено згідно правила', 'checkbox'), $is_customer_note = 0, $added_by_user = false);
-                return;
-            }*/
-
-            /** Check if receipt is already created */
-            if (! empty(get_post_meta($order_id, 'receipt_id', true))) {
-                $order->add_order_note(__('Чек вже створено', 'checkbox'), $is_customer_note = 0, $added_by_user = false);
-                return;
-            }
-
-            /** Check if Autoopen shift feature is activated */
-            if(is_null(get_option('ppo_autoopen_shift'))){
-                update_option( 'ppo_autoopen_shift', 1 );
-            }
-           
-
-            $login       = get_option('ppo_login');
-            $password    = get_option('ppo_password');
-            $cashbox_key = get_option('ppo_cashbox_key');
-
-            if ($login && $password && $cashbox_key) {
-                $is_dev = boolval(get_option('ppo_is_dev_mode'));
-                $logger = new Checkbox\KLoggerDecorator(boolval(get_option('ppo_logger')));
-                $api = new Checkbox\API($login, $password, $cashbox_key, $is_dev);
-
-                $shift = $api->getCurrentCashierShift();
-
-                 if (1 === (int) get_option('ppo_autoopen_shift')) {
-                    if (isset($shift['status']) && ( 'OPENED' === $shift['status'] )) {
-                        $logger->info(__('Зміна вже відкрита', 'checkbox'));
-                    }
-                    else{
-                        $logger->info(__('Зміна не відкрита. Початок відкриття зміни', 'checkbox'));
-                        mrkv_checkbox_connect(false);
-                        sleep(8); // wait for 8 sec while shift is opening
-
-                        if (0 === (int) get_option('ppo_connected')) {
-                            mrkv_checkbox_connect(false);
-                        }
-
-                        $shift = $api->getCurrentCashierShift();
-                    }
-                }
-
-                if (isset($shift['status']) && ( 'OPENED' === $shift['status'] )) {
-                    $logger->info(__('Зміна відкрита. Початок створення чека', 'checkbox'));
-                    $result = mrkv_checkbox_create_receipt($api, $order);
-
-                    if ($result) {
-                        $order->add_order_note(sprintf('%s. <a href="%s" target="_blank">%s</a>', __('Чек створено', 'checkbox'), "https://check.checkbox.ua/{$result['receipt_id']}", __('Роздрукувати', 'checkbox')), $is_customer_note = 0, $added_by_user = false);
-
-                        $logger->info(sprintf('Замовлення №%d. %s.', $order->get_id(), __('Чек створено', 'checkbox')));
-                    } else {
-                        $order->add_order_note(__('Помилка створення чеку', 'checkbox'), $is_customer_note = 0, $added_by_user = false);
-
-                        $logger->error(sprintf('Замовлення №%d. %s.', $order->get_id(), __('Виникла помилка під час створення чека', 'checkbox') . '.'
-                        . __('Повідомлення:', 'checkbox') . ' ' . $result['message'] . '. '
-                        . __('Деталі:', 'checkbox') . " loc:{$result['detail']['loc']}, msg:{$result['detail']['msg']}, type:{$result['detail']['type']}"));
-
-                        $logger->debug(__('Зміст помилки:', 'checkbox'), $result);
-                    }
-                } else {
-                    $logger->info(__('Зміна не відкрита для створення чека', 'checkbox'));
-                    $order->add_order_note(__('Зміна не відкрита', 'checkbox'), $is_customer_note = 0, $added_by_user = false);
-
-                    $logger->error(sprintf('Замовлення №%d. %s.', $order->get_id(), __('Зміна не відкрита', 'checkbox'), $order->get_id()));
-                }
-            }
-        }
-    }
-}
-
-if (! function_exists('mrkv_checkbox_create_receipt')) {
-    /**
-     * Receipt creation
-     *
-     * @param Checkbox\API $api Checkbox API
-     * @param WC_Order         $order Order
-     */
-    function mrkv_checkbox_create_receipt($api, $order)
-    {
-
-        $payment_settings = get_option('ppo_payment_type');
-        $user             = wp_get_current_user();
-
-        $cashier_name = get_option('ppo_cashier_name') . ' ' . get_option('ppo_cashier_surname');
-        // $departament  = 'store';
-
-        $params         = array();
-        $order_data     = $order->get_data();
-        $goods_items    = $order->get_items();
-        $payment_method = $order->get_payment_method();
-
-        $email = isset($order_data['billing']['email']) ? $order_data['billing']['email'] : $user->user_email;
-
-        $payment_type = isset($payment_settings[ $payment_method ]) ? mb_strtoupper($payment_settings[ $payment_method ]) : 'CASHLESS';
-
-        $goods       = array();
-        $total_price = 0;
-        $tax = get_option('ppo_tax_code');
-        /**
-         * @let WC_Order_Item_Product $item
-         */
-        foreach ($goods_items as $item) {
-            $price = ($item->get_subtotal() / $item->get_quantity());
-            $price_checkbox = floatval($price) * 100;
-
-            $good = array(
-                'code'  => $item->get_id() . '-' . $item->get_name(),
-                'name'  => $item->get_name(),
-                'price' => round($price_checkbox),
-            );
-
-            if (!empty($tax)) {
-                $good['tax'] = array($tax);
-            }
-
-            $total_price += $price * $item->get_quantity() * 100;
-
-            $goods[] = array(
-                'good'     => $good,
-                'quantity' => (int) ($item->get_quantity() * 1000),
-            );
-        }
-
-        $params['goods']        = $goods;
-        $params['cashier_name'] = $cashier_name;
-        // $params['departament']  = $departament;
-
-        if (!empty($email)) {
-            $params['delivery'] = array( 'email' => $email );
-        }
-
-        if( count( $order->get_coupon_codes() ) > 0 ) {
-            $discount_type = 'fixed_cart';
-            $coupon_amount = $order->get_total_discount();
-
-            $mode = 'VALUE';
-            $coupon_amount = $coupon_amount * 100;
-
-            $params['discounts'][] = array(
-                'type' => 'DISCOUNT',
-                'mode' => $mode,
-                'value' => $coupon_amount,
-                'name' => 'Купон'
-            );
-            $total_price = $order->get_total() * 100;
-        }
-
-        $params['payments'][]   = array(
-            'type'  => $payment_type,
-            'value' => ceil($total_price),
-        );
-
-        $footer = get_option('ppo_receipt_footer');
-        if ($footer) {
-            $params['footer'] = Checkbox\Support::processReceiptFooter($order, $footer);
-        }
-
-        $result  = array();
-        $receipt = $api->createReceipt($params);
-
-        if (isset($receipt['error'])) {
-            $result['success'] = false;
-            $result['message'] = sprintf('%s: %d.%s: %s.', __('Код помилки', 'checkbox'), $receipt['error']['code'], __('Повідомлення:', 'checkbox'), $receipt['error']['message']);
-
-            return $result;
-        }
-
-        /** Save order ID */
-        if (isset($receipt['id'])) {
-            /** Save receipt ID in meta */
-            update_post_meta($order->get_id(), 'receipt_id', sanitize_text_field($receipt['id']));
-            update_post_meta($order->get_id(), 'receipt_url', sprintf('https://check.checkbox.ua/%s', $receipt['id']));
-            $result['success'] = true;
-            $result['receipt_id'] = $receipt['id'];
-        } else {
-            $result['success'] = false;
-            $result['message'] = $receipt['message'];
-        }
-
-        return $result;
-    }
-}
-
-// -----------------------------------------------------------------------//
-// ---------------------------DASHBOARD WIDGET----------------------------//
-// -----------------------------------------------------------------------//
-
-add_action('wp_dashboard_setup', 'mrkv_checkbox_ppo_status_dashboard_widget');
-if (! function_exists('mrkv_checkbox_ppo_status_dashboard_widget')) {
-    /**
-     * Register plugin dashboard widget only for admin role
-     */
-    function mrkv_checkbox_ppo_status_dashboard_widget()
-    {
-        if (current_user_can('activate_plugins')) {
-            wp_add_dashboard_widget('status_widget', 'Checkbox', 'mrkv_checkbox_status_widget_form');
-        }
-    }
-}
-
-if (! function_exists('mrkv_checkbox_status_widget_form')) {
-    /**
-     * Plugin dashboard widget functionality
-     */
-    function mrkv_checkbox_status_widget_form()
-    {
-        $shift        = '';
-        $shift_id     = '';
-        $is_connected = false;
-        $status       = __('Закрито', 'checkbox');
-
-        $login       = get_option('ppo_login');
-        $password    = get_option('ppo_password');
-        $cashbox_key = get_option('ppo_cashbox_key');
-
-        if ($login && $password && $cashbox_key) {
-            $is_dev = boolval(get_option('ppo_is_dev_mode'));
-            $api   = new Checkbox\API($login, $password, $cashbox_key, $is_dev);
-            $shift = $api->getCurrentCashierShift();
-
-            if (isset($shift['status'])) {
-                $shift_id = $shift['id'] ?? '';
-
-                if ('OPENED' === $shift['status']) {
-                    $is_connected = true;
-                    $status       = __('Відкрито', 'checkbox');
-
-                    if ((int) get_option('ppo_connected') !== 1) {
-                        update_option('ppo_connected', 1);
-                    }
-                } else {
-                    if ((int) get_option('ppo_connected') !== 0) {
-                        update_option('ppo_connected', 0);
-                    }
-                }
-            }
-        }
-
-        ?>
-        <form>
-            <p><?php esc_html_e('Статус', 'checkbox'); ?>: <span id="ppo_status" class="status" style="font-weight: 500; text-transform: uppercase;"><?php echo esc_html($status); ?></span></p>
-            <div class="ppo_connect-group" style="<?php echo esc_html(( $is_connected ) ? 'display: none;' : 'display: inline-flex;'); ?> align-items: center;¨" >
-                <button type="button" id="ppo_button_connect" class="start button button-secondary"><?php esc_html_e('Відкрити зміну', 'checkbox'); ?></button>
-                <img id="ppo_process" style="display: none; margin-left: 10px;" src="<?php echo esc_url(plugins_url('assets/img/ajax-loader.gif', __FILE__)); ?>" width="20px" height="20px" alt="proccess imgage" />
-            </div>
-            <div class="ppo_disconnect-group" style="<?php echo esc_html(( ! $is_connected ) ? 'display: none;' : 'display: inline-flex;'); ?> align-items: center;">
-                <button type="button" id="ppo_button_disconnect" class="end button button-secondary"><?php esc_html_e('Закрити зміну', 'checkbox'); ?></button>
-                <img id="ppo_process" style="display: none; margin-left: 10px;" src="<?php echo esc_url(plugins_url('assets/img/ajax-loader.gif', __FILE__)); ?>" width="20px" height="20px" alt="proccess imgage" />
-            </div>
-
-            <input type="hidden" id="ppo_shift_id" value="<?php echo esc_html($shift_id); ?>">
-        </form>
-
-        <?php
-    }
-}
-
-// -----------------------------------------------------------------------//
-// --------------------------CONNECT AND DISCONNECT-----------------------//
-// -----------------------------------------------------------------------//
-
-add_action('wp_ajax_mrkv_checkbox_check_connection', 'mrkv_checkbox_check_connection');
-if (! function_exists('mrkv_checkbox_check_connection')) {
-    /**
-     * Check connection
-     */
-    function mrkv_checkbox_check_connection()
-    {
-
-        check_ajax_referer('ppo_checkconnect');
-
-        $res = array();
-
-        $login       = get_option('ppo_login');
-        $password    = get_option('ppo_password');
-        $cashbox_key = get_option('ppo_cashbox_key');
-
-        if ($login && $password && $cashbox_key) {
-            $shift_id = isset($_POST['shift_id']) ? sanitize_text_field(wp_unslash($_POST['shift_id'])) : '';
-
-            if ($shift_id) {
-                $is_dev         = boolval(get_option('ppo_is_dev_mode'));
-                $api            = new Checkbox\API($login, $password, $cashbox_key, $is_dev);
-                $response       = $api->checkConnection((int) $shift_id);
-                $status         = $response['status'] ?? '';
-                $res['status']  = $status;
-                $res['message'] = '';
-
-                wp_send_json_success($res);
-
-                return true;
-            } else {
-                wp_send_json_error(
-                    array(
-                        'message' => __('Відсутній ID зміни', 'checkbox'),
-                    )
-                );
-            }
-        } else {
-            wp_send_json_error(
-                array(
-                    'message' => __("Будь ласка, заповніть обов'язкові поля в налаштуваннях плагіна Checkbox", 'checkbox'),
-                )
-            );
-
-            return false;
-        }
-    }
-}
-
-add_action('wp_ajax_mrkv_checkbox_connect', 'mrkv_checkbox_connect');
-if (! function_exists('mrkv_checkbox_connect')) {
-    /**
-     * Shift opening
-     *
-     * !this function is used directly and by WP AJAX
-     */
-    function mrkv_checkbox_connect($this_ajax = true)
-    {
-        if($this_ajax){
-            if (wp_doing_ajax()) {
-            check_ajax_referer('ppo_connect');
-        }
-
-        $res = array();
-
-        $login       = get_option('ppo_login');
-        $password    = get_option('ppo_password');
-        $cashbox_key = get_option('ppo_cashbox_key');
-
-        if ($login && $password && $cashbox_key) {
-            $is_dev = boolval(get_option('ppo_is_dev_mode'));
-            $logger = new Checkbox\KLoggerDecorator(boolval(get_option('ppo_logger')));
-            $api    = new Checkbox\API($login, $password, $cashbox_key, $is_dev);
-
-            $shift = $api->connect();
-
-            if (isset($shift['id'])) {
-                $res['shift_id'] = $shift['id'];
-                $res['status']   = ( 'CREATED' === $shift['status'] ) ? __('Відкрито', 'checkbox') : $shift['status'];
-                $res['message']  = '';
-
-                update_option('ppo_connected', 1);
-
-                $logger->info(__('Зміна успішно відкрита', 'checkbox'));
-
-                if (wp_doing_ajax()) {
-                    wp_send_json_success($res);
-                }
-            } else {
-                $res['shift_id'] = '';
-
-                if ('Not authenticated' === $shift['message']) {
-                    $res['message'] = __('Невірний логін або пароль. Будь ласка, перевірте дані доступу до особистого кабінету касира на сервісі Checkbox.', 'checkbox');
-
-                    $logger->error(__('Невірний логін або пароль. Будь ласка, перевірте дані доступу до особистого кабінету касира на сервісі Checkbox.', 'checkbox'));
-                } else {
-                    $res['message'] = $shift['message'];
-
-                    $logger->error(sprintf('%s %s', __('Під час відкриття зміни виникла помилка. Повідомлення:', 'checkbox'), $shift['message']));
-                }
-
-                if (wp_doing_ajax()) {
-                    wp_send_json_error($res);
-                }
-            }
-        } else {
-            if (wp_doing_ajax()) {
-                wp_send_json_error(
-                    array(
-                        'message' => __("Будь ласка, заповніть обов'язкові поля в налаштуваннях плагіна PPO Checkbox", 'checkbox'),
-                    )
-                );
-            }
-        }
-        }
-        else{
-        $res = array();
-
-        $login       = get_option('ppo_login');
-        $password    = get_option('ppo_password');
-        $cashbox_key = get_option('ppo_cashbox_key');
-
-        if ($login && $password && $cashbox_key) {
-            $is_dev = boolval(get_option('ppo_is_dev_mode'));
-            $logger = new Checkbox\KLoggerDecorator(boolval(get_option('ppo_logger')));
-            $api    = new Checkbox\API($login, $password, $cashbox_key, $is_dev);
-
-            $shift = $api->connect();
-
-            if (isset($shift['id'])) {
-                $res['shift_id'] = $shift['id'];
-                $res['status']   = ( 'CREATED' === $shift['status'] ) ? __('Відкрито', 'checkbox') : $shift['status'];
-                $res['message']  = '';
-
-                update_option('ppo_connected', 1);
-
-                $logger->info(__('Зміна успішно відкрита', 'checkbox'));
-
-                /*if (wp_doing_ajax()) {
-                    wp_send_json_success($res);
-                }*/
-            } else {
-                $res['shift_id'] = '';
-
-                if ('Not authenticated' === $shift['message']) {
-                    $res['message'] = __('Невірний логін або пароль. Будь ласка, перевірте дані доступу до особистого кабінету касира на сервісі Checkbox.', 'checkbox');
-
-                    $logger->error(__('Невірний логін або пароль. Будь ласка, перевірте дані доступу до особистого кабінету касира на сервісі Checkbox.', 'checkbox'));
-                } else {
-                    $res['message'] = $shift['message'];
-
-                    $logger->error(sprintf('%s %s', __('Під час відкриття зміни виникла помилка. Повідомлення:', 'checkbox'), $shift['message']));
-                }
-
-                /*if (wp_doing_ajax()) {
-                    wp_send_json_error($res);
-                }*/
-            }
-        } else {
-            /*if (wp_doing_ajax()) {
-                wp_send_json_error(
-                    array(
-                        'message' => __("Будь ласка, заповніть обов'язкові поля в налаштуваннях плагіна PPO Checkbox", 'checkbox'),
-                    )
-                );
-            }*/
-        }
-        }
-    }
-}
-
-add_action('wp_ajax_mrkv_checkbox_disconnect', 'mrkv_checkbox_disconnect');
-if (! function_exists('mrkv_checkbox_disconnect')) {
-    /**
-     * Shift closing
-     *
-     * !this function is used by CRON and WP AJAX
-     * */
-    function mrkv_checkbox_disconnect()
-    {
-
-        if (wp_doing_ajax()) {
-            check_ajax_referer('ppo_disconnect');
-        }
-
-        $res = array();
-
-        $login       = get_option('ppo_login');
-        $password    = get_option('ppo_password');
-        $cashbox_key = get_option('ppo_cashbox_key');
-
-        if ($login && $password && $cashbox_key) {
-            $is_dev = boolval(get_option('ppo_is_dev_mode'));
-            $logger = new Checkbox\KLoggerDecorator(boolval(get_option('ppo_logger')));
-            $api    = new Checkbox\API($login, $password, $cashbox_key, $is_dev);
-
-            $shift = $api->disconnect();
-
-            if (isset($shift['id'])) {
-                $res['shift_id'] = $shift['id'];
-                $res['status']   = ( 'CLOSING' === $shift['status'] ) ? __('Закрито', 'checkbox') : $shift['status'];
-                $res['message']  = '';
-
-                update_option('ppo_connected', 0);
-
-                $logger->info(__('Зміна успішно закрита', 'checkbox'));
-
-                if (wp_doing_ajax()) {
-                    wp_send_json_success($res);
-                }
-            } else {
-                $res['shift_id'] = '';
-                $res['message']  = $shift['message'];
-
-                $logger->error(sprintf('%s %s', __('Під час закриття зміни виникла помилка. Повідомлення:', 'checkbox'), $shift['message']));
-
-                if (wp_doing_ajax()) {
-                    wp_send_json_error($res);
-                }
-            }
-        } else {
-            if (wp_doing_ajax()) {
-                wp_send_json_error(
-                    array(
-                        'message' => __("Будь ласка, заповніть обов'язкові поля в налаштуваннях плагіна PPO Checkbox", 'checkbox'),
-                    )
-                );
-            }
-        }
-    }
-}
-
-add_action('admin_print_scripts', 'mrkv_checkbox_ppo_connect_script', 999);
-if (! function_exists('mrkv_checkbox_ppo_connect_script')) {
-    /**
-     * Script for plugin dashboard widget
-     */
-    function mrkv_checkbox_ppo_connect_script()
-    {
-
-        if ('dashboard' !== get_current_screen()->base) {
-            return;
-        }
-        ?>
-
-        <script>
-            jQuery(document).ready(function($) {
-
-                $('#ppo_button_connect').on('click', function() {
-
-                    $.ajax({
-                        type: 'post',
-                        url: '<?php echo admin_url('admin-ajax.php'); ?>',
-                        data: {
-                            action: 'mrkv_checkbox_connect',
-                            _wpnonce: '<?php echo wp_create_nonce('ppo_connect'); ?>'
-                        },
-                        dataType: 'json',
-                        beforeSend: function() {
-                            $('.ppo_connect-group img#ppo_process').show();
-                        },
-                        error: function(response) {
-                            $('.ppo_connect-group img#ppo_process').hide();
-
-                            alert( _e('Виникла помилка. Спробуйте ще раз.', 'checkbox') );
-                        },
-                        success: function(response) {
-                            $('.ppo_connect-group img#ppo_process').hide();
-
-                            if (response.success) {
-                                let shift_id = response.data.shift_id;
-                                let status = response.data.status;
-                                $('#ppo_status').html(status);
-                                $('#ppo_shift_id').val(shift_id);
-                                $('.ppo_connect-group').hide();
-                                $('.ppo_disconnect-group').css('display', 'inline-flex');
-                            } else {
-                                alert(response.data.message);
-                            }
-                        }
-                    });
-
-                });
-
-                $('#ppo_button_disconnect').on('click', function() {
-
-                    $.ajax({
-                        type: 'post',
-                        url: ajaxurl,
-                        data: {
-                            action: 'mrkv_checkbox_disconnect',
-                            _wpnonce: '<?php echo wp_create_nonce('ppo_disconnect'); ?>'
-                        },
-                        dataType: 'json',
-                        beforeSend: function() {
-                            $('.ppo_disconnect-group img#ppo_process').show();
-                        },
-                        error: function() {
-                            $('.ppo_disconnect-group img#ppo_process').hide();
-
-                            alert( esc_html_e('Виникла помилка. Спробуйте ще раз.', 'checkbox') );
-                        },
-                        success: function(response) {
-                            $('.ppo_disconnect-group img#ppo_process').hide();
-
-                            if (response.success) {
-                                let status = response.data.status;
-                                $('#ppo_status').html(status);
-                                $('.ppo_disconnect-group').hide();
-                                $('.ppo_connect-group').css('display', 'inline-flex');
-                            } else {
-                                alert(response.data.message);
-                            }
-                        }
-                    });
-
-                });
-
-                let timer = setInterval(() => checkStatus(), 5000);
-
-                function checkStatus() {
-                    let shift_id = $('#ppo_shift_id').val();
-                    if (shift_id) {
-                        let request = $.post(
-                            ajaxurl, {
-                                action: 'mrkv_checkbox_check_connection',
-                                shift_id: shift_id,
-                                _wpnonce: '<?php echo wp_create_nonce('ppo_checkconnect'); ?>'
-                            }
-                        );
-                    }
-                }
-            });
-        </script>
-
-        <?php
-    }
-}
-
-if (! function_exists('mrkv_checkbox_show_plugin_admin_page')) {
-    /** Plugin admin page content */
-    function mrkv_checkbox_show_plugin_admin_page()
-    {
-        mrkv_checkbox_send_request( 'updated' );
-        ?>
-        <style>
-            .table_input {
-                width: 100%
-            }
-
-            .table_textarea {
-                width: 100%;
-            }
-
-            span.tooltip {
-                background-image: url('<?php echo plugin_dir_url(__FILE__) ?>assets/img/tooltip-icon.svg');
-            }
-
-            .table_storagesList {
-                width: 200px
-            }
-
-            .gateway-settings__wrapper {
-                width: 100%;
-                height: 100%;
-                overflow-x: auto;
-            }
-
-            table.gateway-settings {
-                width: 100%;
-                border-collapse: collapse;
-            }
-
-            table.gateway-settings thead {
-                white-space: nowrap;
-            }
-
-            table.gateway-settings th {
-                background-color: #bbb;
-            }
-
-            table.gateway-settings th,
-            table.gateway-settings td {
-                border: 1px solid #555;
-                text-align: center;
-                padding-right: 30px;
-            }
-
-            table.gateway-settings .gateway-title {
-                max-width: 250px;
-            }
-
-            table.gateway-settings .gateway-title p {
-                text-align: left;
-                text-overflow: ellipsis;
-                white-space: nowrap;
-                overflow: hidden;
-                width: 100%;
-            }
-
-            table.gateway-settings tbody td label:not(:last-child) {
-                margin-right: 7px;
-            }
-            .wrap-checkbox-setting{
-                background: #fff;
-                padding: 20px 40px;
-                margin-top: 20px;
-                overflow: hidden;
-            }
-            .checkbox-setting-row{
-                display: flex;
-                align-items: flex-start;
-                flex-wrap: wrap;
-                justify-content: space-between;
-            }
-            .checkbox-setting-col{
-                width: calc(50% - 40px);
-            }
-            .wrap-checkbox-setting > h2{
-                display: flex;
-                align-items: center;
-            }
-            .wrap-checkbox-setting > h2 img{
-                width: 14px;
-                margin-left: 10px;
-                background: #000;
-                padding: 4px;
-                border-radius: 100%;
-            }
-            .checkbox-setting-data-title{
-                font-weight: 600;
-                display: flex;
-                align-items: center;
-            }
-            .checkbox-setting-data input[type="checkbox"]{
-                display: none;
-            }
-            .checkbox-setting-data-meta{
-                display: inline-block;
-                width: fit-content;
-                padding-right: 15px;
-            }
-            .checkbox-setting-data-meta input{
-                display: none;
-            }
-            .checkbox-setting-data-meta label{
-                position: relative;
-                display: flex;
-                align-items: center;
-            }
-            .checkbox-setting-data-meta label:before{
-                content: '';
-                position: relative;
-                display: inline-block;
-                vertical-align: middle;
-                height: 20px;
-                width: 20px;
-                border-color: #000;
-                border-style: solid;
-                border-width: 2px;
-                box-sizing: border-box;
-                border-radius: 50%;
-                margin-right: 10px;
-                -webkit-transition: all 0.2s;
-                transition: all 0.2s;
-                opacity: .6;
-            }
-            .checkbox-setting-data-meta input:checked + label:before{
-                opacity: 1;
-            }
-            .checkbox-setting-data-meta input:checked + label:after{
-                content: '';
-                width: 10px;
-                height: 10px;
-                background: #000;
-                border-radius: 50%;
-                position: absolute;
-                top: 5px;
-                bottom: 0;
-                left: 5px;
-                right: 0;
-                margin: 0;
-                opacity: 1;
-                -webkit-transition: all 0.2s;
-                transition: all 0.2s;
-            }
-            .checkbox-setting-data:has(input[type="radio"]){
-                margin-bottom: 18px;
-            }
-            .mrkv_table-payment__body__checkbox__input {
-                position: relative;
-                width: 50px;
-                height: 25px;
-                display: inline-block;
-                margin-right: 5px;
-                min-width: 28px;
-            }
-            .mrkv_checkbox_slider {
-                position: absolute;
-                cursor: pointer;
-                top: 0;
-                left: 0;
-                right: 0;
-                bottom: 0;
-                background-color: #ccc;
-                -webkit-transition: .4s;
-                transition: .4s;
-                border-radius: 34px;
-            }
-            .mrkv_checkbox_slider:before{
-                position: absolute;
-                content: "";
-                height: 21px;
-                width: 21px;
-                left: 2px;
-                bottom: 2px;
-                background-color: white;
-                -webkit-transition: .4s;
-                transition: .4s;
-                border-radius: 50%;
-            }
-            input[type="checkbox"]:checked + label .mrkv_checkbox_slider:before{
-                -webkit-transform: translateX(24px);
-                -ms-transform: translateX(24px);
-                transform: translateX(24px);
-            }
-            input[type="checkbox"]:checked + label .mrkv_checkbox_slider{
-                background-color: #8f257d;
-            }
-            .checkbox-setting-data input[type="checkbox"]{
-                display: none;
-            }
-            .checkbox-setting-data .chosen-container{
-                width: 100% !important;
-            }
-            .checkbox-setting-data .chosen-choices{
-                box-shadow: unset;
-                border-radius: 4px;
-                border: 1px solid #8c8f94 !important;
-                box-shadow: unset !important;
-                background-image: unset;
-            }
-            .checkbox-setting-col input[type="submit"]{
-                background-color: #8f257d;
-                border-radius: 12px;
-                height: 40px;
-                font-size: 14px;
-                line-height: 20px;
-                display: flex;
-                align-items: center;
-                justify-content: center;
-                min-width: 152px;
-                color: #fff;
-                border: unset;
-                font-weight: 500;
-            }
-            .checkbox-setting-col input[type="submit"]:hover{
-                background-color: #8f257dab;
-            }
-            .checkbox-setting-col .gateway-settings{
-                border: unset;
-            }
-            .checkbox-setting-col .gateway-settings th,
-            .checkbox-setting-col .gateway-settings td{
-                border: unset;
-                border-bottom: 2px solid #dcdcde;
-                text-align: left;
-            }
-            .checkbox-setting-col .gateway-settings th{
-                background-color: transparent;  
-                font-size: 13px;
-                margin-bottom: 0;
-                font-weight: 500;
-            }
-            .mt-40 {
-                margin-top: 40px;
-            }
-            .plugin-development span {
-                margin-bottom: 6px;
-                font-size: 16px;
-                display: block;
-                font-weight: 500;
-            }
-            .plugin-development img {
-                height: 27px;
-            }
-            /*.checkbox-setting-col-8{
-                width: calc(75% - 40px);
-            }
-            .checkbox-setting-col-4{
-                    width: calc(30% - 40px);
-            }*/
-            .checkbox-setting-col .gateway-settings .chosen-choices{
-                box-shadow: unset;
-                border-radius: 4px;
-                border: 1px solid #8c8f94 !important;
-                box-shadow: unset !important;
-                background-image: unset;
-            }
-            .checkbox-setting-col .gateway-settings input[type="radio"] + label:before{
-                content: '';
-                position: relative;
-                display: inline-block;
-                vertical-align: middle;
-                height: 20px;
-                width: 20px;
-                border-color: #000;
-                border-style: solid;
-                border-width: 2px;
-                box-sizing: border-box;
-                border-radius: 50%;
-                margin-right: 10px;
-                -webkit-transition: all 0.2s;
-                transition: all 0.2s;
-                opacity: .6;
-            }
-            .checkbox-setting-col .gateway-settings input[type="radio"]{
-                display: none;
-            }
-            .checkbox-setting-col .gateway-settings input[type="radio"] + label{
-                position: relative;
-                display: flex;
-                align-items: center;
-                margin-top: 5px;
-                margin-bottom: 5px;
-            }
-            .checkbox-setting-col .gateway-settings input[type="radio"]:checked + label:before{
-                opacity: 1;
-            }
-            .checkbox-setting-col .gateway-settings input[type="radio"]:checked + label:after{
-                content: '';
-                width: 10px;
-                height: 10px;
-                background: #000;
-                border-radius: 50%;
-                position: absolute;
-                top: 5px;
-                bottom: 0;
-                left: 5px;
-                right: 0;
-                margin: 0;
-                opacity: 1;
-                -webkit-transition: all 0.2s;
-                transition: all 0.2s;
-            }
-            .checkbox-log-pre{
-                margin-top: 20px;
-                border: 1px solid #dcdcde;
-                padding: 20px;
-                background: #fbfbfb;
-                height: 300px;
-                overflow-y: scroll;
-            }
-            .btn-call-clean-log{
-                cursor: pointer;
-                background-color: #00000029;
-                border-radius: 12px;
-                height: 40px;
-                font-size: 14px;
-                line-height: 20px;
-                display: flex;
-                align-items: center;
-                justify-content: center;
-                min-width: 152px;
-                color: #000;
-                border: unset;
-                font-weight: 500;
-                width: fit-content;
-            }
-            .btn-call-clean-log:hover{
-                opacity: .7;
-            }
-
-            @media (max-width: 782px) {
-                table.gateway-settings th, table.gateway-settings td {
-                    display: block;
-                    justify-content: center;
-                    align-items: center;
-                    width: auto;
-                    height: 50px;
-                    white-space: normal;
-                }
-                table.gateway-settings tbody tr {
-                    white-space: nowrap;
-                }
-                div.gateway-settings__wrapper {
-                    overflow-x: auto;
-                }
-                .checkbox-setting-col, .checkbox-setting-col-8, .checkbox-setting-col-4{
-                    width: 100%;
-                }
-                .wrap-checkbox-setting{
-                    padding: 20px 15px; 
-                }
-            }
-
-            @media (max-width: 700px) {
-                div.gateway-settings__wrapper {
-                    overflow-x: auto;
-                    width: 100%;
-                }
-            }
-        </style>
-        <div class="wrap wrap-checkbox-setting">
-            <h2>
-                <?php echo get_admin_page_title(); ?>
-                <img src="<?php echo plugin_dir_url(__FILE__) ?>assets/img/checkbox-icon-setting.jpg" alt="Checkbox">
-                </h2>
-                <hr>
-                <p><?php esc_html_e('Checkbox – програмний РРО, створений для бізнесу будь-якого масштабу. Це сучасний сервіс, який здешевить реєстрацію розрахункових операцій, а також спростить і впорядкує ведення звітності та контроль платежів.', 'checkbox'); ?></p>
-
-            <?php settings_errors(); ?>
-
-            <form method="post" action="options.php">
-                <?php settings_fields('ppo-settings-group'); ?>
-
-                <div class="checkbox-setting-row">
-                    <div class="checkbox-setting-col">
-                        <h2><?php esc_html_e('Основна інформація про касира', 'checkbox'); ?></h2>
-                        <hr>
-                        <div class="checkbox-setting-data">
-                            <p class="checkbox-setting-data-title"><?php esc_html_e('Логін касира', 'checkbox'); ?></p>
-                            <input class="table_input" type="text" name="ppo_login" value="<?php echo esc_html(get_option('ppo_login')); ?>" />
-                        </div>
-                        <div class="checkbox-setting-data">
-                            <p class="checkbox-setting-data-title"><?php esc_html_e('Особистий пароль касира', 'checkbox'); ?></p>
-                            <input class="table_input" type="password" name="ppo_password" value="<?php echo esc_html(get_option('ppo_password')); ?>" />
-                        </div>
-                        <div class="checkbox-setting-data">
-                            <p class="checkbox-setting-data-title"><?php esc_html_e('Ім\'я касира', 'checkbox'); ?></p>
-                            <input class="table_input" type="text" name="ppo_cashier_name" value="<?php echo esc_html(get_option('ppo_cashier_name')); ?>" />
-                        </div>
-                        <div class="checkbox-setting-data">
-                            <p class="checkbox-setting-data-title"><?php esc_html_e('Прізвище касира', 'checkbox'); ?></p>
-                            <input class="table_input" type="text" name="ppo_cashier_surname" value="<?php echo esc_html(get_option('ppo_cashier_surname')); ?>" />
-                        </div>
-                    </div>
-                    <div class="checkbox-setting-col">
-                        <h2><?php esc_html_e('Налаштування каси', 'checkbox'); ?></h2>
-                        <hr>
-                    <div class="checkbox-setting-data">
-                        <p class="checkbox-setting-data-title"><?php esc_html_e('Ключ каси', 'checkbox'); ?></p>
-                        <input class="table_input" type="text" name="ppo_cashbox_key" value="<?php echo esc_html(get_option('ppo_cashbox_key')); ?>" />
-                    </div>
-                    <div class="checkbox-setting-data">
-                        <p class="checkbox-setting-data-title"><?php esc_html_e('ЄДРПОУ', 'checkbox'); ?></p>
-                        <input class="table_input" type="text" name="ppo_cashbox_edrpou" value="<?php echo esc_html(get_option('ppo_cashbox_edrpou')); ?>" />
-                    </div>
-                        <div class="checkbox-setting-data">
-                            <p class="checkbox-setting-data-title"><?php esc_html_e('Код податку', 'checkbox'); ?></p>
-                            <?php 
-                                $all_payment_code = array('', '8');
-                            ?>
-                            <select name="ppo_tax_code" id="ppo_tax_code">
-                                <?php 
-                                    foreach($all_payment_code as $code){
-                                        $selected = '';
-                                        if($code == get_option('ppo_tax_code')){
-                                            $selected = 'selected';
-                                        }
-                                        ?>
-                                            <option <?php echo $selected; ?> value="<?php echo $code; ?>"><?php echo $code; ?></option>        
-                                        <?php
-                                    }
-                                ?>
-                            </select>
-                        </div>
-                    <div class="checkbox-setting-data">
-                        <p class="checkbox-setting-data-title"><?php esc_html_e('Спосіб підпису', 'checkbox'); ?> <span class="tooltip" aria-label="<?php echo esc_html('Доступні два механізми підпису чеків: Checkbox Підпис — утиліта, що встановлюється на будь-якому комп’ютері з доступом до Інтернету, і HSM, або Checkbox Cloud, — сертифікований хмарний сервіс для генерації та зберігання ключів DepositSign, у разі вибору якого необхідність встановлення будь-якого ПЗ для роботи з ЕЦП відсутня.', 'checkbox'); ?>" data-microtip-position="right" role="tooltip"></span></p>
-                         <?php
-                            $ppo_sign_method = get_option('ppo_sign_method');
-                        ?>
-                        <div class="checkbox-setting-data-meta">
-                            <input class="table_input" type="radio" name="ppo_sign_method" id="ppo_sign_method_cloud" value="cloud"
-                            <?php
-                            if (isset($ppo_sign_method)) {
-                                checked($ppo_sign_method, 'cloud');
-                            }
-                            ?>
-                             />
-                            <label for="ppo_sign_method_cloud"><?php esc_html_e('Checkbox Cloud', 'checkbox'); ?></label>
-                        </div>
-                        <div class="checkbox-setting-data-meta">
-                            <input class="table_input" type="radio" name="ppo_sign_method" id="ppo_sign_method_soft" value="soft"
-                            <?php
-                            if (isset($ppo_sign_method)) {
-                                checked($ppo_sign_method, 'soft');
-                            }
-                            ?>
-                             />
-                            <label for="ppo_sign_method_soft"><?php esc_html_e('Checkbox Підпис', 'checkbox'); ?></label>
-                        </div>
-                    </div>
-                    <div class="checkbox-setting-data">
-                        <p class="checkbox-setting-data-title"><?php esc_html_e('Відкривати зміну при створенні чека', 'checkbox'); ?> <span class="tooltip" aria-label="<?php echo esc_html('Зміна відкриватиметься автоматично при створенні чека.', 'checkbox'); ?>" data-microtip-position="right" role="tooltip"></span></p>
-                        <input class="table_input" type="checkbox" name="ppo_autoopen_shift" id="ppo_autoopen_shift" value="1" <?php checked(get_option('ppo_autoopen_shift'), 1); ?> />
-                        <label for="ppo_autoopen_shift">
-                            <div class="mrkv_table-payment__body__checkbox__input">
-                                <span class="mrkv_checkbox_slider"></span>
-                            </div>
-                        </label>
-                    </div>
-                </div>
-            </div>
-            <div class="checkbox-setting-row">
-                <div class="checkbox-setting-col checkbox-setting-col-8">
-                    <h2><?php esc_html_e('Налаштування автоматичного створення чеків', 'checkbox'); ?></h2>
-                    <hr>
-                    <?php update_option( 'ppo_autoopen_shift', 1 ); ?>
-                    <div class="checkbox-setting-data" style="display: none;">
-                        <p class="checkbox-setting-data-title">
-                            <?php esc_html_e("Автоматично створювати чеки", 'checkbox'); ?> <span class="tooltip" aria-label="<?php echo esc_html('Чек створюватиметься автоматично при зміні статусу замовлення. При ввімкненому стані в табличці "Налаштування способів оплати" з\'явиться колонка "Ігнорувати чек?", в якій ви можете для кожного способу оплати дозволити або заборонити автоматичне створення чеку при зміні статусу.', 'checkbox'); ?>" data-microtip-position="right" role="tooltip"></span>
-                        </p>
-                        <input class="table_input" type="checkbox" name="ppo_autocreate" id="ppo_autocreate" value="1" <?php checked(get_option('ppo_autocreate'), 1); ?> />
-                        <label for="ppo_autocreate">
-                            <div class="mrkv_table-payment__body__checkbox__input">
-                                <span class="mrkv_checkbox_slider"></span>
-                            </div>
-                        </label>
-                    </div>
-                    <div class="order-statuses" style="<?php echo ( 1 == get_option('ppo_autocreate') ) ? '' : 'display: none;'; ?>">
-                        <div class="checkbox-setting-data">
-                            <p class="checkbox-setting-data-title">
-                                <?php esc_html_e("Статуси замовлення", 'checkbox'); ?> <span style="color: red;">(<?php esc_html_e("застаріле", 'checkbox'); ?>)</span>
-                            <span class="tooltip" aria-label="<?php echo esc_html('Виберіть статуси замовлення, при зміні на які буде створюватися чек.', 'checkbox'); ?>" data-microtip-position="right" role="tooltip">
-                                </span>
-                            </p>
-                            <?php
-                                $all_order_statuses = wc_get_order_statuses();
-                                $autocreate_receipt_statuses = (array) get_option('ppo_autocreate_receipt_order_statuses');
-                            ?>
-                            <select class="chosen order-statuses" name="ppo_autocreate_receipt_order_statuses[]" data-placeholder="<?php _e('Виберіть статуси замовлення', 'checkbox') ?>" multiple>
-                                <?php
-                                if (! empty($all_order_statuses)) :
-                                    foreach ($all_order_statuses as $k => $v) :
-                                        $k = str_replace('wc-', '', $k);
-                                        if (empty($autocreate_receipt_statuses)) :
-                                            ?>
-                                        <option value="<?php echo $k; ?>"><?php echo $v; ?></option>
-                                            <?php
-                                        else :
-                                            ?>
-                                        <option value="<?php echo $k; ?>" <?php echo ( in_array($k, $autocreate_receipt_statuses) ) ? 'selected' : ''; ?>><?php echo $v; ?></option>
-                                            <?php
-                                        endif;
-                                    endforeach;
-                                else :
-                                    printf('<option value="">%s</option>', __('None'));
-                                endif;
-                                ?>
-                            </select>
-                        </div>
-                    </div>
-                        <div class="checkbox-setting-data">
-                            <p class="checkbox-setting-data-title">
-                                <?php esc_html_e('Правила автоматичного формування чеків', 'checkbox'); ?> <span class="tooltip" aria-label="<?php echo esc_html('Визначення типу для кожного способу оплати необхідне для створення чека.', 'checkbox'); ?>" data-microtip-position="right" role="tooltip"></span>
-                            </p>
-                        </div>
-                         <?php
-                            $enabled_gateways = array_filter(WC()->payment_gateways->payment_gateways(), function ($gateway) {
-                                return 'yes' === $gateway->enabled;
-                            });
-
-                            $ppo_autocreate = get_option('ppo_autocreate');
-                            $ppo_autocreate_payment_order_statuses = get_option('ppo_autocreate_payment_order_statuses');
-                            $ppo_payment_type          = get_option('ppo_payment_type');
-                            $ppo_skip_receipt_creation = get_option('ppo_skip_receipt_creation');
-                            ?>
-                            <div class="gateway-settings__wrapper">
-                                <table class="gateway-settings">
-                                    <thead>
-                                        <tr>
-                                            <th><?php esc_html_e('Спосіб оплати', 'checkbox'); ?></th>
-                                            <th><?php esc_html_e('Тип', 'checkbox'); ?></th>
-                                            <th class="select-order-statuses" style="<?php echo (1 !== (int) $ppo_autocreate) ? 'display:none;' : ''; ?>"><?php esc_html_e('Статуси замовлення', 'checkbox'); ?></th>
-                                        </tr>
-                                    </thead>
-                                    <tbody>
-                                        <?php
-                                        foreach ($enabled_gateways as $id => $gateway) :
-                                            ?>
-                                            <tr>
-                                                <td class="gateway-title" title="<?php echo esc_html($gateway->get_title()); ?>">
-                                                    <p><?php echo esc_html($gateway->get_title()); ?></p>
-                                                </td>
-                                                <td>
-                                                    <input type="radio" name="ppo_payment_type[<?php echo esc_html($id); ?>]" id="ppo_payment_type_cash[<?php echo esc_html($id); ?>]" <?php
-                                                    if (isset($ppo_payment_type[$id])) {
-                                                        checked($ppo_payment_type[$id], 'cash');
-                                                    }
-                                                    ?> value="cash">
-                                                    <label for="ppo_payment_type_cash[<?php echo esc_html($id); ?>]">CASH</label>
-
-                                                    <input type="radio" name="ppo_payment_type[<?php echo esc_html($id); ?>]" id="ppo_payment_type_cashless[<?php echo esc_html($id); ?>]" <?php
-                                                    if (isset($ppo_payment_type[$id])) {
-                                                        checked($ppo_payment_type[$id], 'cashless');
-                                                    }
-                                                    ?> value="cashless">
-                                                    <label for="ppo_payment_type_cashless[<?php echo esc_html($id); ?>]">CASHLESS</label>
-                                                </td>
-                                                <td class="select-order-statuses" style="<?php echo (1 !== (int) $ppo_autocreate) ? 'display:none;' : ''; ?>">
-                                                    <select class="chosen order-statuses" name="ppo_autocreate_payment_order_statuses[<?php echo esc_html($id); ?>][]" data-placeholder="<?php _e('Виберіть статуси замовлення', 'checkbox') ?>" multiple>
-                                                        <?php
-                                                        if (! empty($all_order_statuses)) :
-                                                            foreach ($all_order_statuses as $k => $v) :
-                                                                $k = str_replace('wc-', '', $k);
-                                                                ?>
-                                                            <option value="<?php echo $k; ?>" <?php echo ( isset($ppo_autocreate_payment_order_statuses[$id]) && in_array($k, $ppo_autocreate_payment_order_statuses[$id]) ) ? 'selected' : ''; ?>><?php echo $v; ?></option>
-                                                                <?php
-                                                            endforeach;
-                                                        else :
-                                                            printf('<option value="">%s</option>', __('None'));
-                                                        endif;
-                                                        ?>
-                                                    </select>
-                                                </td>
-                                            </tr>
-                                            <?php
-                                        endforeach;
-                                        ?>
-                                    </tbody>
-                                </table>
-                            </div>
-                </div>
-                <div class="checkbox-setting-col checkbox-setting-col-4">
-                    <h2><?php esc_html_e('Додаткові налаштування', 'checkbox'); ?></h2>
-                    <hr>
-                    <div class="checkbox-setting-data">
-                        <p class="checkbox-setting-data-title">
-                            <?php esc_html_e("Службова інформація", 'checkbox'); ?> <span class="tooltip" aria-label="<?php echo esc_html('Значення данного поля буде відображатися в нижній частині електронного чеку.', 'checkbox'); ?>" data-microtip-position="right" role="tooltip"></span>
-                        </p>
-                        <textarea class="table_textarea" type="text" name="ppo_receipt_footer"/><?php echo get_option('ppo_receipt_footer'); ?></textarea>
-                        <br>
-                        <span style="color: grey;">Доступні шорткоди: [website_title], [order_id], [order_created_date], [order_paid_date].</span>
-                    </div>
-                    <div class="checkbox-setting-data">
-                        <p class="checkbox-setting-data-title">
-                            <?php esc_html_e("Тестовий режим", 'checkbox'); ?>
-                        <span class="tooltip" aria-label="<?php echo esc_html('При ввімкненому тестовому режимі, всі запити будуть спрямовані до тестового сервера Checkbox — dev-api.checkbox.in.ua. Для підключення ви повинні ввести "Логін", "Пароль" і "Ліцензійний ключ ВКА" від тестового акаунта. Тестовий акаунт надається за проханням адміністрацією Checkbox.', 'checkbox'); ?>" data-microtip-position="right" role="tooltip"></span>
-                        </p>
-                        <input disabled="" class="table_input" type="checkbox" name="ppo_is_dev_mode" id="ppo_is_dev_mode" value="1" <?php checked(get_option('ppo_is_dev_mode'), 1); ?> />
-                        <label for="ppo_is_dev_mode">
-                            <div class="mrkv_table-payment__body__checkbox__input">
-                                <span class="mrkv_checkbox_slider"></span>
-                            </div>
-                        </label>
-                            <br>
-                            <span style="color: grey;">Тестовий режим не використовується, якщо ключ каси починається на "test".</span>
-                    </div>
-                    <div class="checkbox-setting-data">
-                        <p class="checkbox-setting-data-title">
-                            <?php esc_html_e("Ввімкнути логування", 'checkbox'); ?> <span class="tooltip" aria-label="<?php echo esc_html('Використовуйте лише для логування помилок.', 'checkbox'); ?>" data-microtip-position="right" role="tooltip"></span>
-                        </p>
-                        <input class="table_input" type="checkbox" name="ppo_logger" id="ppo_logger" value="1" <?php checked(get_option('ppo_logger'), 1); ?> />
-                        <label for="ppo_logger">
-                            <div class="mrkv_table-payment__body__checkbox__input">
-                                <span class="mrkv_checkbox_slider"></span>
-                            </div>
-                        </label>
-                    </div>
-                    <div class="checkbox-setting-data">
-                        <p class="checkbox-setting-data-title">
-                            <?php esc_html_e("Лог", 'checkbox'); ?>
-                        </p>
-                        <?php 
-                            $log_file = file_get_contents(plugin_dir_url(__FILE__) . "logs/checkbox.log");
-                        ?>
-                        <pre class="checkbox-log-pre"><?php echo $log_file; ?></pre>
-                        <div class="btn-call-clean-log"><?php echo __('Очистити лог', 'checkbox'); ?></div>
-                    </div>
-                    <?php echo submit_button(__('Зберегти', 'checkbox')); ?>
-                </div>
-            </div>
-            <div class="plugin-development mt-40">
-                <span>Веб студія</span>
-                <a href="https://morkva.co.ua/" target="_blank"><img src="<?php echo plugin_dir_url(__FILE__) ?>assets/img/morkva-logo.svg" alt="Morkva" title="Morkva"></a>
-            </div>
-            </form>
-        </div>
-        <script>
-            jQuery(function($){
-                jQuery('.btn-call-clean-log').click(function(){
-                    jQuery.ajax({
-                        url: '<?php echo admin_url( "admin-ajax.php" ) ?>',
-                        type: 'POST',
-                        data: 'action=checkbox_clean_log', 
-                        success: function( data ) {
-                            jQuery('.checkbox-log-pre').text('');
-                        }
-                    });
-                });
-            });
-        </script>
-        <?php
-    }
-}
-
-add_action( 'wp_ajax_checkbox_clean_log', 'checkbox_clean_log_func' );
-add_action( 'wp_ajax_nopriv_checkbox_clean_log', 'checkbox_clean_log_func' );
- 
-function checkbox_clean_log_func(){
-    file_put_contents( plugin_dir_path(__FILE__) . "logs/checkbox.log", '');
-
-    die; 
-}
-
-// -----------------------------------------------------------------------//
-// ----------------------------ADDITIONAL---------------------------------//
-// -----------------------------------------------------------------------//
-
-add_action('admin_enqueue_scripts', 'mrkv_checkbox_styles_and_scripts');
-if (! function_exists('mrkv_checkbox_styles_and_scripts')) {
-
-    function mrkv_checkbox_styles_and_scripts($hook)
-    {
-        if ('toplevel_page_checkbox_settings' != $hook) {
-            return;
-        }
-        // custom style and script
-        wp_enqueue_style('checkbox', plugin_dir_url(__FILE__) . 'assets/css/checkbox.css', array(), CHECKBOX_VERSION);
-        wp_enqueue_script('checkbox', plugin_dir_url(__FILE__) . 'assets/js/checkbox.js', array('jquery'), CHECKBOX_VERSION, true);
-        // microtip
-        wp_enqueue_style('checkbox-microtip', plugin_dir_url(__FILE__) . 'assets/css/microtip.min.css');
-        // chosen
-        wp_enqueue_style('checkbox-chosen', plugin_dir_url(__FILE__) . 'assets/css/chosen.min.css', array(), '1.8.7');
-        wp_enqueue_script('checkbox-chosen', plugin_dir_url(__FILE__) . 'assets/js/chosen.jquery.min.js', array('jquery'), '1.8.7', true);
-    }
-
-}
 
 // -----------------------------------------------------------------------//
 // ------------------ACTIVATION AND DEACTIVATION HOOKS--------------------//
 // -----------------------------------------------------------------------//
 
-function mrkv_checkbox_send_request(string $status)
-{
-    $ip       = !empty($_SERVER['SERVER_ADDR']) ? $_SERVER['SERVER_ADDR'] : $_SERVER['REMOTE_ADDR'];
-    $home_url = parse_url(home_url());
+# Include all needed classes for activation and deactivation plugin
+require_once 'includes/class-activation-deactivation.php'; 
 
-    $data = [
-        'ip'      => $ip,
-        'domain'  => $home_url['host'],
-        'product' => 'checkbox',
-        'version' => CHECKBOX_VERSION,
-        'license' => CHECKBOX_LICENSE,
-        'info'    => get_option('ppo_cashbox_key'),
-        'status'  => $status
-    ];
-    
-    $url = 'https://api2.morkva.co.ua/api/customers/register';
+# Setup activation and deactivation plugin
+new MRKV_ACTIVATION_DEACTIVATION();
 
-    $response = wp_remote_post( $url, array(
-        'method'      => 'POST',
-        'timeout'     => 45,
-        'redirection' => 5,
-        'blocking'    => true,
-        'headers'     => array(
-            'Accept' => 'application/json'
-        ),
-        'body'        => $data,
-        'cookies'     => array()
-    ));
-        
-    if ( is_wp_error( $response ) ) {
-        $error_message = $response->get_error_message();
-        
-        $logger = new Checkbox\KLoggerDecorator(boolval(get_option('ppo_logger')));
-        $logger->error( "Something went wrong: $error_message", 1 );
-    }
-}
+// -----------------------------------------------------------------------//
+// --------------------------------SETUP----------------------------------//
+// -----------------------------------------------------------------------//
 
-// send request when plugin's settings are being saved
-add_filter('pre_update_option_ppo_cashbox_key', function ( $value, $old_value, $option ) {
-    mrkv_checkbox_send_request( 'updated' );
+# Include all needed classes for setup
+require_once 'includes/class-setup.php'; 
 
-    return $value;
-}, 10, 3);
+# Setup plugin
+new MRKV_CHECKBOX_SETUP(__FILE__);
 
-add_action('upgrader_process_complete', 'mrkv_checkbox_upgrade', 10, 2);
-function mrkv_checkbox_upgrade($upgrader_object, $options)
-{
-    $current_plugin_path_name = plugin_basename(__FILE__);
+// -----------------------------------------------------------------------//
+// -----------------------WOOCOMMERCE CUSTOMISATION-----------------------//
+// -----------------------------------------------------------------------//
 
-    if ($options['action'] == 'update' && $options['type'] == 'plugin') {
-        foreach ($options['plugins'] as $each_plugin) {
-            if ($each_plugin == $current_plugin_path_name) {
-                mrkv_checkbox_send_request( 'updated' );
-            }
-        }
-    }
-}
+# Include all needed classes for woocommerce
+require_once 'includes/class-woocommerce.php'; 
 
-register_activation_hook(__FILE__, 'mrkv_checkbox_activation_cb');
-function mrkv_checkbox_activation_cb()
-{
-    update_option( 'ppo_autoopen_shift', 1 );
+# Setup woocommerce settings
+new MRKV_CHECKBOX_WOOCOMMERCE();
 
-    if (wp_next_scheduled('checkbox_close_shift')) {
-        wp_clear_scheduled_hook('checkbox_close_shift');
-    }
+// -----------------------------------------------------------------------//
+// -----------------------------CREATE RECEIPT----------------------------//
+// -----------------------------------------------------------------------//
 
-    if (wp_next_scheduled('checkbox_open_shift')) {
-        wp_clear_scheduled_hook('checkbox_open_shift');
-    }
+# Include all needed classes for create receipt
+require_once 'includes/class-create-receipt.php'; 
 
-    mrkv_checkbox_send_request( 'activated' );
-}
+# Setup create receipt
+new MRKV_CHECKBOX_RECEIPT();
 
-register_deactivation_hook(__FILE__, 'mrkv_checkbox_deactivation_cb');
-function mrkv_checkbox_deactivation_cb()
-{
-    if (get_option('ppo_connected')) {
-        mrkv_checkbox_disconnect();
-    }
+// -----------------------------------------------------------------------//
+// ---------------------------DASHBOARD WIDGET----------------------------//
+// -----------------------------------------------------------------------//
 
-    mrkv_checkbox_send_request( 'deactivated' );
-}
+# Include all needed classes for dashboard settings
+require_once 'includes/class-dashboard-widget.php'; 
+
+# Setup dashboard settings
+new MRKV_DASHBOARD_WIDGET(__FILE__);
+
+// -----------------------------------------------------------------------//
+// --------------------------CONNECT AND DISCONNECT-----------------------//
+// -----------------------------------------------------------------------//
+
+# Include all needed classes for conntect disconnect shift
+require_once 'includes/class-conntect-disconnect.php'; 
+
+# Setup conntect disconnect shift
+new MRKV_CONNTECT_DISCONNECT();
+
